blockhouse/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.2.2,<6.0
django-cors-headers>=3.14.0,<4.0
